+++ src/modules/sbstudio/plugin/operators/export_to_drotek.py
@@ -40,9 +40,6 @@ class DrotekExportOperator(ExportOperator):
     )
 
     def get_format(self) -> FileFormat:
-        """Returns the file format that the operator uses. Must be overridden
-        in subclasses.
-        """
         return FileFormat.DROTEK
 
     def get_operator_name(self) -> str: