+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added mandatory backend version check before dispatching the first request to
   the current server instance.
 
-### Changed 
+### Changed
 
 - The minimum backend version required for this version of the add-on is now
   2.29.0.